Initial Webpack 4 Upgrade
