+++ package.json
@@ -134,9 +134,9 @@
     "velocity-animate": "^1.5.1",
     "vinyl-buffer": "^1.0.1",
     "vinyl-source-stream": "^2.0.0",
-    "webpack": "^3.11.0",
-    "webpack-dev-server": "^2.11.2",
-    "webpack-manifest-plugin": "^1.3.2",
+    "webpack": "^4.17.2",
+    "webpack-dev-server": "3.1.8",
+    "webpack-manifest-plugin": "2.0.0-rc.2",
     "wnumb": "^1.1.0"
   },
   "scripts": {
@@ -159,5 +159,8 @@
   },
   "engines": {
     "yarn": ">= 1.0.0"
+  },
+  "devDependencies": {
+    "webpack-cli": "^3.1.0"
   }
 }
